% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.to.delta.R
\name{acc.to.delta}
\alias{acc.to.delta}
\title{Calculates delta = log(d') given accuracy or accuracy given delta
assuming no bias and p(stim = 1) = 0.5.}
\usage{
acc.to.delta(acc = 0.75, delta = NULL, crit = 0)
}
\arguments{
\item{acc}{a vector of average accuracy scores. All the values must
be greater than 0 and lower than 1.}

\item{delta}{default is NULL, if provided the delta values will be
converted to accuracy of an unbiased observer}
}
\value{
delta = log(d') or expected accuracy
}
\description{
Calculates delta = log(d') given accuracy or accuracy given delta
assuming no bias and p(stim = 1) = 0.5.
}
