% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhsdtr.R
\name{bhsdtr}
\alias{bhsdtr}
\title{main interface}
\usage{
bhsdtr(
  model.formulae,
  response.formula,
  data,
  links = list(gamma = "log_distance"),
  fit = "ml",
  thresholds_scale = 2,
  iter = 20000,
  stanargs = NULL
)
}
\description{
main interface
}
