% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.R
\name{cumulative}
\alias{cumulative}
\title{Fit a cumulative ordinal model with varying ordered thresholds and fixed latent mean and sd}
\usage{
cumulative(formula, ...)
}
\arguments{
\item{formula}{a model formula of the form resp ~ f1 * f2 + (f1 |
id), where resp is the name of the ordinal response
variable. This will be transformed to a vector of two model
formulae and a response - stimulus formula for the bhsdtr
function: bhsdtr(c(mean ~ 1, thr ~ f1 * f2 + (f1 | id)), resp
~ 1, ...)}

\item{...}{additional arguments passed to the bhsdtr function. See
the bhsdtr documentation for for details.}
}
\value{
a bhsdtr model object
}
\description{
Fit a cumulative ordinal model with varying ordered thresholds and fixed latent mean and sd
}
