% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{fit a bhsdtr model using ml or stan}
\usage{
fit(
  m,
  method = "ml",
  ml_init = 0,
  chains = parallel::detectCores() - 1,
  iter = max(20000/chains, 5000),
  warmup = 2000,
  init_r = 0.5,
  stan_optimizations = T,
  ...
)
}
\description{
fit a bhsdtr model using ml or stan
}
