% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.names.R
\name{predictor.names}
\alias{predictor.names}
\title{Extract the names of the predictor variables}
\usage{
predictor.names(model, ...)
}
\arguments{
\item{model}{a bhsdtr model object}

\item{...}{e.g., 'delta', 'gamma'}
}
\value{
vector of names of variables in the original data frame
}
\description{
Extract the names of the predictor variables
}
