% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bhsdtr_samples}
\alias{print.bhsdtr_samples}
\title{print bhsdtr condition-specific samples or ml point estimates}
\usage{
\method{print}{bhsdtr_samples}(x, digits = 2, probs = c(0.025, 0.975), ...)
}
\description{
print bhsdtr condition-specific samples or ml point estimates
}
