% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples}
\alias{samples}
\title{Obtain posterior samples or ml estimates for unique combinations of predictor values}
\usage{
samples(m, par, group = NULL, ...)
}
\arguments{
\item{m}{a fitted bhsdtr model object}

\item{par}{the name of a parameter. Can be either "dprim",
"metad", "thr", "sdratio", or "mean".}

\item{group}{grouping variable number. For example, if this is the
d' formula: dprim ~ f1 * f2 + (f1 | id) + (f2 | id), then
group = 1 corresponds to the (f1 | id) part, and group = 2
corresponds to the (f2 | id) part. If this is not NULL then
you will get samples for every unique combination of
predictors and levels of this grouping factor.}
}
\value{
a S x D x C array, where S is the number of posterior
    samples (= 1 for bhsdtr models fitted using ML), D is the
    dimensionality of the parameter (dprim = 1, metad = 2, thr = K
    - 1, mean = 1, sdratio = 1), and C is the number of unique
    combinations of values of predictors for the given parameter.
}
\description{
Obtain posterior samples or ml estimates for unique combinations of predictor values
}
