% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.hit.fa.R
\name{sdt.hit.fa}
\alias{sdt.hit.fa}
\title{Utility function}
\usage{
sdt.hit.fa(dprim, crit)
}
\arguments{
\item{dprim}{d'}

\item{crit}{main decision criterion centered at the midpoint between the evidence distributions}
}
\description{
Utility function
}
