% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.sdt.R
\name{sim.sdt}
\alias{sim.sdt}
\title{Simulate the data from an SDT model}
\usage{
sim.sdt(
  n = 1,
  dprim = 1.5,
  criteria = c(-2.1, -1.4, -0.7, 0, 0.7, 1.4, 2.1),
  sd_ratio = 1
)
}
\arguments{
\item{n}{how many samples to produce}

\item{dprim}{d'}

\item{sd_ratio}{standard deviation ratio}

\item{crit}{main decision criteria centered at the midpoint between the evidence distributions}
}
\description{
Simulate the data from an SDT model
}
