% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bhsdtr_samples}
\alias{summary.bhsdtr_samples}
\title{summary of condition specific samples}
\usage{
\method{summary}{bhsdtr_samples}(x, digits = 1, probs = c(0.025, 0.975), ...)
}
\description{
summary of condition specific samples
}
