% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased.R
\name{unbiased}
\alias{unbiased}
\title{Calculates "unbiased" thresholds for the parsimonious link function}
\usage{
unbiased(K)
}
\description{
Calculates "unbiased" thresholds for the parsimonious link function
}
