% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhsdtr.R
\name{bhsdtr}
\alias{bhsdtr}
\title{main interface}
\usage{
bhsdtr(
  model_formulae,
  response_formula,
  data,
  links = list(gamma = "log_distance"),
  fit_method = "ml",
  thresholds_scale = 2,
  ...
)
}
\description{
main interface
}
