% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhsdtr.R
\name{bhsdtr}
\alias{bhsdtr}
\title{main interface}
\usage{
bhsdtr(
  model_formulae,
  response_formula,
  data,
  links = list(gamma = "log_distance"),
  fit_method = "ml",
  thresholds_scale = 2,
  ...
)
}
\arguments{
\item{model_formulae}{a vector of lmer-style model formulae for
the parameters of the model}

\item{response_formula}{a formula specifying the names of the
response variable and the stimulus variable}

\item{links}{a list specyfing the link functions}

\item{fit_method}{either 'ml' or 'stan'}

\item{thresholds_scale}{thresholds scaling factor (see the bhsdtr preprint)}
}
\value{
a bhsdtr_model object
}
\description{
main interface
}
