% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.specific.samples.R
\name{condition.specific.samples}
\alias{condition.specific.samples}
\title{get posterior samples or jmap point estimates per condition}
\usage{
condition.specific.samples(
  m,
  par,
  group = NULL,
  method = NULL,
  include.vars = NULL
)
}
\description{
get posterior samples or jmap point estimates per condition
}
