% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bhsdtr_model.R
\name{plot.bhsdtr_model}
\alias{plot.bhsdtr_model}
\title{plot a fitted bhsdtr model}
\usage{
\method{plot}{bhsdtr_model}(
  x,
  vs = NULL,
  type = "response",
  alpha = 0.05,
  bw = FALSE,
  verbose = T,
  ...
)
}
\description{
plot a fitted bhsdtr model
}
