% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.prior.R
\name{set.prior}
\alias{set.prior}
\title{Set the parameters of the prior distributions}
\usage{
set.prior(model, ...)
}
\arguments{
\item{model}{a bhsdtr model object}

\item{...}{e.g., delta_prior_fixed_mu = log(1.5),
delta_prior_fixed_sd = 2. Note that the prior parameters are
internally represented by matrices of dimension DxC, where D
is the dimensionality of the parameter (delta has 2 in the
meta-d' model, otherwise it has 1, ) and C is the number of
columns in the fixed (<par>_prior_fixed_mu,
<par>_prior_fixed_sd) or the random (<par>prior_scale_<g>)
effects' model matrix. You can provide scalars, vectors, or
matrices. A vector will fill the prior parameter matrix in
column major order.}
}
\value{
a bhsdtr model object with updated priors
}
\description{
Set the parameters of the prior distributions
}
